+++ src/components/index.tsx
@@ -3,6 +3,7 @@ import Footer from './footer';
 import MaxWidthContent from './max_width_content';
 import MaxWidthContainer from './max_width_container';
 import DataBlocksHeader from './data_blocks_header';
+import StabilitiesItem from './stabilities';
 
 export {
   Layout,
@@ -10,4 +11,5 @@ export {
   MaxWidthContent,
   MaxWidthContainer,
   DataBlocksHeader,
+  StabilitiesItem,
 };
